+++ package.json
@@ -8,10 +8,10 @@
 	],
 	"version": "2.3.1",
 	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build && touch build/.nojekyll",
+		"dev": "vite dev",
+		"build": "vite build && touch build/.nojekyll",
 		"package": "pnpm run test && svelte-kit package && node clean-package.js",
-		"preview": "svelte-kit preview",
+		"preview": "vite preview",
 		"check": "svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. .",
@@ -24,45 +24,51 @@
 		"deploy": "pnpm run deploy:docs && pnpm run deploy:package"
 	},
 	"devDependencies": {
-		"@babel/preset-env": "^7.16.7",
-		"@changesets/cli": "^2.20.0",
-		"@sveltejs/adapter-auto": "next",
-		"@sveltejs/adapter-static": "^1.0.0-next.24",
-		"@sveltejs/kit": "next",
-		"@testing-library/jest-dom": "^5.16.1",
-		"@testing-library/svelte": "^3.0.3",
-		"@types/jest": "^27.4.0",
-		"autoprefixer": "^10.4.2",
-		"cssnano": "^5.0.15",
-		"jest": "^27.4.7",
+		"@babel/preset-env": "^7.20.2",
+		"@changesets/cli": "^2.26.0",
+		"@sveltejs/adapter-auto": "^1.0.0",
+		"@sveltejs/adapter-static": "^1.0.1",
+		"@sveltejs/kit": "^1.0.11",
+		"@testing-library/jest-dom": "^5.16.5",
+		"@testing-library/svelte": "^3.2.2",
+		"@types/jest": "^27.5.2",
+		"@types/lodash": "^4.14.191",
+		"@types/lodash.kebabcase": "^4.1.7",
+		"@types/markdown-it": "^12.2.3",
+		"@types/prismjs": "^1.26.0",
+		"@types/ua-parser-js": "^0.7.36",
+		"autoprefixer": "^10.4.13",
+		"cssnano": "^5.1.14",
+		"jest": "^27.5.1",
 		"jest-in-case": "^1.0.2",
 		"jest-mock-promise": "^1.1.12",
-		"markdown-it": "^12.3.1",
-		"markdown-it-anchor": "^8.4.1",
+		"markdown-it": "^12.3.2",
+		"markdown-it-anchor": "^8.6.6",
 		"node-sass": "^6.0.1",
-		"postcss": "^8.4.5",
-		"postcss-load-config": "^3.1.1",
-		"prettier": "^2.5.1",
-		"prettier-plugin-svelte": "^2.5.1",
-		"svelte": "^3.45.0",
-		"svelte-check": "^2.2.11",
-		"svelte-jester": "^2.1.5",
-		"svelte-preprocess": "^4.10.1",
-		"svelte2tsx": "^0.4.12",
+		"postcss": "^8.4.21",
+		"postcss-load-config": "^3.1.4",
+		"prettier": "^2.8.2",
+		"prettier-plugin-svelte": "^2.9.0",
+		"svelte": "^3.55.1",
+		"svelte-check": "^2.10.3",
+		"svelte-jester": "^2.3.2",
+		"svelte-preprocess": "^4.10.7",
+		"svelte2tsx": "^0.4.14",
 		"tailwindcss": "^2.2.19",
-		"ts-jest": "^27.1.2",
-		"tslib": "^2.3.1",
-		"typescript": "^4.5.4",
-		"vite-plugin-markdown": "^2.0.2"
+		"ts-jest": "^27.1.5",
+		"tslib": "^2.4.1",
+		"typescript": "^4.9.4",
+		"vite": "^4.0.4",
+		"vite-plugin-markdown": "^2.1.0"
 	},
 	"type": "module",
 	"dependencies": {
 		"is-promise": "^4.0.0",
 		"lodash.kebabcase": "^4.1.1",
 		"prism-svelte": "^0.4.7",
-		"prismjs": "^1.26.0",
-		"svelte-awesome": "^2.4.2",
-		"ua-parser-js": "^1.0.2"
+		"prismjs": "^1.29.0",
+		"svelte-awesome": "^2.4.6",
+		"ua-parser-js": "^1.0.32"
 	},
 	"homepage": "https://chainlist.github.io/svelte-forms/",
 	"repository": {
