+++ src/app.html
@@ -1,14 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="description" content="" />
-		<link rel="icon" href="/svelte-forms/favicon.png" />
-		<!-- <link rel="stylesheet" href="/prism.css" /> -->
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		%svelte.head%
-	</head>
-	<body class="text-2xl relative">
-		<div id="svelte">%svelte.body%</div>
-	</body>
+
+<head>
+	<meta charset="utf-8" />
+	<meta name="description" content="" />
+	<link rel="icon" href="/svelte-forms/favicon.png" />
+	<!-- <link rel="stylesheet" href="/prism.css" /> -->
+	<meta name="viewport" content="width=device-width, initial-scale=1" />
+	%sveltekit.head%
+</head>
+
+<body class="text-2xl relative">
+	<div id="svelte">%sveltekit.body%</div>
+</body>
+
 </html>
